Update syncbranches workflows
Use `@pytorchmergebot` credentials to do the merge
Infer sync branch name from the workflow rather than hardcode it
Move common scripts to gitutils.py
